Remove deprecated Navigation nodes
Removes deprecated Navigation / Navigation2D nodes.
