--- doc/classes/Navigation2D.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<class name="Navigation2D" inherits="Node2D" version="3.6" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="../class.xsd">
-	<brief_description>
-		2D navigation and pathfinding node.
-	</brief_description>
-	<description>
-		[i]Deprecated.[/i] [Navigation2D] node and [method get_simple_path] are deprecated and will be removed in a future version. Use [method Navigation2DServer.map_get_path] instead.
-		Navigation2D provides navigation and pathfinding within a 2D area, specified as a collection of [NavigationPolygon] resources. By default, these are automatically collected from child [NavigationPolygonInstance] nodes.
-	</description>
-	<tutorials>
-		<link title="2D Navigation Demo">https://godotengine.org/asset-library/asset/117</link>
-	</tutorials>
-	<methods>
-		<method name="get_closest_point" qualifiers="const">
-			<return type="Vector2" />
-			<argument index="0" name="to_point" type="Vector2" />
-			<description>
-				Returns the navigation point closest to the point given. Points are in local coordinate space.
-			</description>
-		</method>
-		<method name="get_closest_point_owner" qualifiers="const">
-			<return type="RID" />
-			<argument index="0" name="to_point" type="Vector2" />
-			<description>
-				Returns the owner of the [NavigationPolygon] which contains the navigation point closest to the point given. This is usually a [NavigationPolygonInstance].
-			</description>
-		</method>
-		<method name="get_rid" qualifiers="const">
-			<return type="RID" />
-			<description>
-				Returns the object's [RID].
-			</description>
-		</method>
-		<method name="get_simple_path" qualifiers="const">
-			<return type="PoolVector2Array" />
-			<argument index="0" name="start" type="Vector2" />
-			<argument index="1" name="end" type="Vector2" />
-			<argument index="2" name="optimize" type="bool" default="true" />
-			<description>
-				[i]Deprecated.[/i] [Navigation2D] node and [method get_simple_path] are deprecated and will be removed in a future version. Use [method Navigation2DServer.map_get_path] instead.
-				Returns the path between two given points. Points are in local coordinate space. If [code]optimize[/code] is [code]true[/code] (the default), the path is smoothed by merging path segments where possible.
-			</description>
-		</method>
-	</methods>
-	<members>
-		<member name="cell_size" type="float" setter="set_cell_size" getter="get_cell_size" default="1.0">
-			The XY plane cell size to use for fields.
-		</member>
-		<member name="edge_connection_margin" type="float" setter="set_edge_connection_margin" getter="get_edge_connection_margin" default="1.0">
-			This value is used to detect the near edges to connect compatible regions.
-		</member>
-		<member name="navigation_layers" type="int" setter="set_navigation_layers" getter="get_navigation_layers" default="1">
-			A bitfield determining all navigation map layers the navigation can use on a [method Navigation2D.get_simple_path] path query.
-		</member>
-	</members>
-	<constants>
-	</constants>
-</class>
