+++ editor/plugins/tile_set_editor_plugin.cpp
@@ -34,6 +34,7 @@
 #include "core/os/keyboard.h"
 #include "editor/editor_scale.h"
 #include "editor/plugins/canvas_item_editor_plugin.h"
+#include "scene/2d/navigation_polygon_instance.h"
 #include "scene/2d/physics_body_2d.h"
 #include "scene/2d/sprite.h"
