+++ scene/2d/navigation_obstacle_2d.h
@@ -34,15 +34,13 @@
 #include "scene/2d/node_2d.h"
 #include "scene/main/node.h"
 
-class Navigation2D;
-
 class NavigationObstacle2D : public Node {
 	GDCLASS(NavigationObstacle2D, Node);
 
-	Navigation2D *navigation;
 	Node2D *parent_node2d = nullptr;
 	RID agent;
 	RID map_before_pause;
+	RID map_override;
 
 	bool estimate_radius = true;
 	real_t radius = 1.0;
@@ -56,17 +54,14 @@ public:
 	NavigationObstacle2D();
 	virtual ~NavigationObstacle2D();
 
-	void set_navigation(Navigation2D *p_nav);
-	const Navigation2D *get_navigation() const {
-		return navigation;
-	}
-
-	void set_navigation_node(Node *p_nav);
-	Node *get_navigation_node() const;
-
 	RID get_rid() const {
 		return agent;
 	}
+
+	void set_agent_parent(Node *p_agent_parent);
+
+	void set_navigation_map(RID p_navigation_map);
+	RID get_navigation_map() const;
 
 	void set_estimate_radius(bool p_estimate_radius);
 	bool is_radius_estimated() const {
