+++ scene/3d/navigation_obstacle.h
@@ -34,16 +34,14 @@
 #include "scene/3d/spatial.h"
 #include "scene/main/node.h"
 
-class Navigation;
-
 class NavigationObstacle : public Node {
 	GDCLASS(NavigationObstacle, Node);
 
 	Spatial *parent_spatial = nullptr;
-	Navigation *navigation;
 
 	RID agent;
 	RID map_before_pause;
+	RID map_override;
 
 	bool estimate_radius = true;
 	real_t radius = 1.0;
@@ -57,17 +55,14 @@ public:
 	NavigationObstacle();
 	virtual ~NavigationObstacle();
 
-	void set_navigation(Navigation *p_nav);
-	const Navigation *get_navigation() const {
-		return navigation;
-	}
-
-	void set_navigation_node(Node *p_nav);
-	Node *get_navigation_node() const;
-
 	RID get_rid() const {
 		return agent;
 	}
+
+	void set_agent_parent(Node *p_agent_parent);
+
+	void set_navigation_map(RID p_navigation_map);
+	RID get_navigation_map() const;
 
 	void set_estimate_radius(bool p_estimate_radius);
 	bool is_radius_estimated() const {
