+++ scene/resources/tile_set.h
@@ -34,8 +34,8 @@
 #include "core/array.h"
 #include "core/resource.h"
 #include "scene/2d/light_occluder_2d.h"
-#include "scene/2d/navigation_polygon.h"
 #include "scene/resources/convex_polygon_shape_2d.h"
+#include "scene/resources/navigation_polygon.h"
 #include "scene/resources/shape_2d.h"
 #include "scene/resources/texture.h"
