+++ servers/navigation_2d_server.h
@@ -33,7 +33,7 @@
 
 #include "core/object.h"
 #include "core/rid.h"
-#include "scene/2d/navigation_polygon.h"
+#include "scene/resources/navigation_polygon.h"
 
 // This server exposes the 3D `NavigationServer` features in the 2D world.
 class Navigation2DServer : public Object {
